+++ modules/kinesis/variables.tf
@@ -15,7 +15,7 @@ variable "custom_url" {
 }
 
 variable "private_key" {
-  description = "The Coralogix private key which is used to validate your authenticity"
+  description = "The Coralogix Data API key which is used to validate your authenticity"
   type        = string
   sensitive   = true
 }
@@ -80,7 +80,7 @@ variable "tags" {
 }
 
 variable "layer_arn" {
-  description = "In case you are using SM This is the ARN of the Coralogix Security Layer."
+  description = "In case you are using Secret Manager This is the ARN of the Coralogix Security lambda Layer."
   type        = string
   default     = ""
 }
@@ -92,7 +92,7 @@ variable "custom_s3_bucket" {
 }
 
 variable "create_secret" {
-  description = "Set to False In case you want to use SM with your secret that contains coralogix private key"
+  description = "Set to False In case you want to use secrets manager with a predefine secret that was already created and contains Coralogix Send Your Data API key"
   type        = string
   default     = "True"
 }