+++ modules/s3/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## S3
 
 ### 0.0.3 / 1.10.2023
-* [Change] Change SSM option in the integration to SM - Secret Manager.
+* [Change] Change SSM option in the integration to Secret Manager.
 
 ### 0.0.2 / 16.8.2023
 * [Update] Add an option to use an existing secret instead of creating a new one with SSM, and remove ssm_enabled variable.